improve key import and signature verification
Set TOFU specifically for keys we imported and not all (which should
be the same set anyway since we started off a fresh db, but this
may defend against a race accessing the db).
We also  keep track of whether a signature matches a key we imported
(which should already be the case anyway, but belt and braces).